build: fix build warning.
Guenther
(cherry picked from commit 62fcad0c0548ab53e9c20cfd6301972c68172b95)
(cherry picked from commit 69e93e2b88c880f770b416b6f5645b8d4c8bf5d8)

+++ source/lib/smbldap.c
@@ -692,7 +692,7 @@ int smb_ldap_setup_conn(LDAP **ldap_struct, const char *uri)
 		rc = ldap_set_option(*ldap_struct, LDAP_OPT_NETWORK_TIMEOUT, &ct);
 		if (rc != LDAP_SUCCESS) {
 			DEBUG(0,("Failed to setup an ldap connection timeout %d: %s\n",
-				ct.tv_sec, ldap_err2string(rc)));
+				(int)ct.tv_sec, ldap_err2string(rc)));
 		}
 	}
 #endif
